refactor: remove FluentPropertyMapperBase, inline into MqttFluentPropertyMapper
The base class had only one consumer. Each connector's fluent mapper is
self-contained with its own ConcurrentDictionary and interface
implementations, so the shared base added indirection without value.
