+++ src/Namotion.Interceptor.Mqtt/Mapping/MqttFluentPropertyMapper.cs
@@ -1,4 +1,6 @@
 using System;
+using System.Collections.Concurrent;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq.Expressions;
 using System.Threading;
 using System.Threading.Tasks;
@@ -7,22 +9,33 @@ using Namotion.Interceptor.Registry.Abstractions;
 
 namespace Namotion.Interceptor.Mqtt.Mapping;
 
-/// <summary>
-/// Fluent property mapper for configuring MQTT topic mappings using lambda expressions.
-/// </summary>
-/// <typeparam name="TSubject">The subject type whose properties are being mapped.</typeparam>
 public class MqttFluentPropertyMapper<TSubject>
-    : FluentPropertyMapperBase<TSubject, MqttPropertyMapping>,
-      IReversePropertyMapper<MqttPropertyMapping, MqttLookupKey>
+    : IReversePropertyMapper<MqttPropertyMapping, MqttLookupKey>
 {
+    private readonly ConcurrentDictionary<string, MqttPropertyMapping> _mappings = new();
+
     public MqttFluentPropertyMapper<TSubject> Map<TValue>(
         Expression<Func<TSubject, TValue>> selector,
         Action<MqttFluentMappingBuilder> configure)
     {
         var builder = new MqttFluentMappingBuilder();
         configure(builder);
-        SetMapping(selector, builder.Build());
+        _mappings[PropertyPathHelper.GetPathFromExpression(selector.Body)] = builder.Build();
         return this;
+    }
+
+    public bool TryGetMapping(
+        RegisteredSubjectProperty property,
+        [NotNullWhen(true)] out MqttPropertyMapping? mapping)
+    {
+        var path = PropertyPathHelper.GetPathFromProperty(property);
+        if (_mappings.TryGetValue(path, out var stored))
+        {
+            mapping = stored;
+            return true;
+        }
+        mapping = null;
+        return false;
     }
 
     public ValueTask<RegisteredSubjectProperty?> TryGetPropertyAsync(
